backend/requirements.txt
adrf==0.1.12
amqp==5.3.1
anyio==4.12.0
asgiref==3.8.1
astroid==3.3.8
async-property==0.2.2
bcrypt==4.3.0
billiard==4.2.4
Brotli==1.1.0
celery==5.6.2
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==3.1.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cron_descriptor==2.0.6
cryptography==44.0.1
defusedxml==0.7.1
dill==0.3.9
Django==5.2.4
django-celery-beat==2.8.1
django-cors-headers==3.10.1
django-redis==6.0.0
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
dnspython==2.2.1
drf-yasg==1.20.0
et-xmlfile==1.1.0
gevent==25.9.1
greenlet==3.3.0
gunicorn==23.0.0
h11==0.16.0
httptools==0.7.1
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.2
kombu==5.6.2
MarkupSafe==2.1.2
mccabe==0.7.0
mongoengine==0.29.1
mypy-extensions==1.0.0
numpy==2.4.1
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.0
pandas==2.3.3
pillow==12.1.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.52
pycparser==2.22
PyJWT==1.7.1
pylint==3.3.3
pymongo==4.15.5
pytest==8.3.4
pytest-django==4.9.0
pytest-mock==3.14.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2023.2
PyYAML==6.0.3
redis==6.4.0
reportlab==4.4.7
requests==2.28.2
requests-oauthlib==2.0.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.3
stripe==12.4.0
tenacity==9.0.0
tomlkit==0.13.2
typing_extensions==4.14.1
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.40.0
uvicorn-worker==0.4.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.3
whitenoise==6.9.0
zope.event==6.1
zope.interface==8.2
